refactor: 동아리 정보, 소개 수정, 폴더 구조 수정
