+++ src/main/java/com/USWCicrcleLink/server/admin/controller/AdminController.java
@@ -2,7 +2,7 @@ package com.USWCicrcleLink.server.admin.controller;
 
 import com.USWCicrcleLink.server.admin.service.AdminService;
 import com.USWCicrcleLink.server.club.domain.Club;
-import com.USWCicrcleLink.server.club.domain.Leader;
+import com.USWCicrcleLink.server.clubLeader.domain.Leader;
 import com.USWCicrcleLink.server.global.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
